+++ src/isla/solver.py
@@ -375,7 +375,7 @@ class ISLaSolver:
 
         while self.queue:
             # import dill as pickle
-            # state_hash = -5722477832052327252
+            # state_hash = 9107154106757938105
             # out_file = "/tmp/saved_debug_state"
             # if hash(self.queue[0][1]) == state_hash:
             #     with open(out_file, 'wb') as debug_state_file: